+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- XvdFile: Allow fetching dataUnit from hashtables during CryptSectionXts
 - General: Start using CHANGELOG.md
 - XVDTool: For XVC decryption, iterate through all loaded CIKs to find a matching one